updates based on pr review
