+++ protocol-designer/src/components/steplist/__tests__/ModuleStepItems.test.js
@@ -2,12 +2,13 @@
 import React from 'react'
 import { render, shallow, mount } from 'enzyme'
 import { ModuleStepItems } from '../ModuleStepItems'
+import { MAGDECK } from '../../../constants'
 
 let props
 beforeEach(() => {
   props = {
     action: 'action',
-    module: 'magdeck',
+    module: MAGDECK,
     actionText: 'engage',
     labwareDisplayName: 'magnet module',
     labwareNickname: 'maggie',
